jit: Fix the alignment of the function now that I understand it
For the pointer we return it needs to look like a valid OOP. The
trick that was used to align the code to make that true. Otherwise
empty_context_stack will set a DUMMY_NATIVE_IP and we will resume
at the ipOffset (e.g. the beginning of a loop) and not after the
interrupt check.

This fixes something like this:

 a := nil. p := [a := 3] newProcess.
 [p isTerminated] whileFalse: [p singleStep; suspend]